+++ public/js/i18n/en_US.js
@@ -237,6 +237,10 @@ gc2i18n = {
         "Update SQL": "Update SQL",
         "Ignore": "Ignore",
         "Content": "Content",
-        "No tags": "No tags"
+        "No tags": "No tags",
+
+        "You are no longer logged in to GC2. Close or refresh your browser": "You are no longer logged in to GC2. Close or refresh your browser",
+        "You have started Admin for another schema. Either close this here or refresh your browser": "You have started Admin for another schema. Either close this here or refresh your browser",
+        "Something went wrong. Try refreshing your browser": "Something went wrong. Try refreshing your browser"
     }
 };